+++ source/standalone/workflows/robomimic/collect_demonstrations.py
@@ -140,12 +140,14 @@ def main():
                 collector_interface.add(f"obs/{key}", value)
             # -- actions
             collector_interface.add("actions", actions)
+
             # perform action on environment
             obs_dict, rewards, terminated, truncated, info = env.step(actions)
             dones = terminated | truncated
             # check that simulation is stopped or not
             if env.unwrapped.sim.is_stopped():
                 break
+
             # robomimic only cares about policy observations
             # store signals from the environment
             # -- next_obs
@@ -162,6 +164,10 @@ def main():
             # flush data from collector for successful environments
             reset_env_ids = dones.nonzero(as_tuple=False).squeeze(-1)
             collector_interface.flush(reset_env_ids)
+
+            # check if enough data is collected
+            if collector_interface.is_stopped():
+                break
 
     # close the simulator
     collector_interface.close()